+++ apps/meteor/client/components/FingerprintChangeModalConfirmation.tsx
@@ -1,19 +1,20 @@
 import { Box } from '@rocket.chat/fuselage';
+import { GenericModal } from '@rocket.chat/ui-client';
 import DOMPurify from 'dompurify';
 import type { ReactElement } from 'react';
 import { useTranslation } from 'react-i18next';
 
-import GenericModal from './GenericModal';
-
 type FingerprintChangeModalConfirmationProps = {
 	onConfirm: () => void;
 	onCancel: () => void;
+	onClose: () => void;
 	newWorkspace: boolean;
 };
 
 const FingerprintChangeModalConfirmation = ({
 	onConfirm,
 	onCancel,
+	onClose,
 	newWorkspace,
 }: FingerprintChangeModalConfirmationProps): ReactElement => {
 	const { t } = useTranslation();
@@ -25,6 +26,7 @@ const FingerprintChangeModalConfirmation = ({
 			onCancel={onCancel}
 			cancelText={t('Back')}
 			confirmText={newWorkspace ? t('Confirm_new_workspace') : t('Confirm_configuration_update')}
+			onClose={onClose}
 		>
 			<Box
 				is='p'