+++ apps/meteor/client/views/admin/deviceManagement/DeviceManagementAdminPage.tsx
@@ -1,4 +1,4 @@
-import { useRouteParameter } from '@rocket.chat/ui-contexts';
+import { useRouteParameter, useRouter } from '@rocket.chat/ui-contexts';
 import type { ReactElement } from 'react';
 import { useRef } from 'react';
 import { useTranslation } from 'react-i18next';
@@ -10,6 +10,7 @@ import { Page, PageHeader, PageContent } from '../../../components/Page';
 
 const DeviceManagementAdminPage = (): ReactElement => {
 	const { t } = useTranslation();
+	const router = useRouter();
 	const context = useRouteParameter('context');
 	const deviceId = useRouteParameter('id');
 
@@ -24,7 +25,7 @@ const DeviceManagementAdminPage = (): ReactElement => {
 				</PageContent>
 			</Page>
 			{context === 'info' && deviceId && (
-				<ContextualbarDialog>
+				<ContextualbarDialog onClose={() => router.navigate('/admin/device-management')}>
 					<DeviceManagementInfo deviceId={deviceId} onReload={reloadRef.current} />
 				</ContextualbarDialog>
 			)}