+++ apps/meteor/client/views/omnichannel/contactInfo/ContactInfoError.tsx
@@ -6,20 +6,21 @@ import {
 	ContextualbarTitle,
 	ContextualbarClose,
 	ContextualbarEmptyContent,
+	ContextualbarDialog,
 } from '../../../components/Contextualbar';
 
 const ContactInfoError = ({ onClose }: { onClose: () => void }) => {
 	const { t } = useTranslation();
 
 	return (
-		<>
+		<ContextualbarDialog onClose={onClose}>
 			<ContextualbarHeader>
 				<ContextualbarIcon name='user' />
 				<ContextualbarTitle>{t('Contact')}</ContextualbarTitle>
 				<ContextualbarClose onClick={onClose} />
 			</ContextualbarHeader>
 			<ContextualbarEmptyContent icon='user' title={t('Contact_not_found')} />
-		</>
+		</ContextualbarDialog>
 	);
 };