+++ apps/meteor/server/lib/i18n.ts
@@ -1,4 +1,5 @@
-import i18nDict, { availableTranslationNamespaces, defaultTranslationNamespace, extractTranslationNamespaces } from '@rocket.chat/i18n';
+import { availableTranslationNamespaces, defaultTranslationNamespace, extractTranslationNamespaces } from '@rocket.chat/i18n';
+import i18nDict from '@rocket.chat/i18n/dist/resources';
 
 import { i18n } from '../../app/utils/lib/i18n';